+++ Glyph.h
@@ -5,8 +5,8 @@ struct Glyph
 {
 	unsigned int sprite_num;
 	sf::Color color;
-	sf::Vector2i position;
+	sf::Vector2f position;
 
-	Glyph(unsigned int sprite_num, sf::Color color, sf::Vector2i position)  
+	Glyph(unsigned int sprite_num, sf::Color color, sf::Vector2f position)  
 		: sprite_num(sprite_num), color(color), position(position) {}
 };
