+++ ViewportSystem.h
@@ -11,8 +11,8 @@ private:
 	unsigned int threshold = { 8 };
 	unsigned int height;
 	unsigned int width;
-	sf::Vector2i origin = { -1, -1 };
-	sf::Vector2i old_origin;
+	sf::Vector2f origin = { -1.00f, -1.00f };
+	sf::Vector2f old_origin;
 	std::list<Entity*> on_screen_entities;
 	float updaterate = 30.0;
 	float dt_count = 0.0;
